fix per @tqchen's comments
